Move common logic for hooks into another module
